Add the id to question and option

+++ domain/src/main/java/com/roacult/kero/oxxy/domain/modules/Option.kt
@@ -1,3 +1,3 @@
 package com.roacult.kero.oxxy.domain.modules
 
-data class Option (val id :Long ,val option :String)
+data class Option (val id :Int ,val option :String)
